Various tree refactoring.
