+++ src/main/java/rtg/world/gen/feature/tree/rtg/TreeRTG.java
@@ -1,4 +1,4 @@
-package rtg.world.gen.feature.tree;
+package rtg.world.gen.feature.tree.rtg;
 
 import java.util.Random;
