+++ src/main/java/rtg/world/gen/feature/tree/rtg/pinaceae/TreeRTGPinaceaeBase.java
@@ -5,7 +5,7 @@ import java.util.Random;
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.world.World;
-import rtg.world.gen.feature.tree.TreeRTG;
+import rtg.world.gen.feature.tree.rtg.TreeRTG;
 
 public class TreeRTGPinaceaeBase extends TreeRTG
 {
